refactor(gwt): extract board view helpers from main
